+++ todoman/cli.py
@@ -75,6 +75,10 @@ def _validate_start_date_param(ctx, param, val):
         raise click.BadParameter(e)
 
 
+def _sort_callback(ctx, param, val):
+    return val.split(',') if val else []
+
+
 def _todo_property_options(command):
     click.option(
         '--due', '-d', default='', callback=_validate_date_param,
@@ -211,6 +215,8 @@ def edit(ctx, id, todo_properties, interactive):
     '''
     database = ctx.obj['db']
     todo = database.todo(id)
+    old_list = todo.list
+
     changes = False
     for key, value in todo_properties.items():
         if value:
@@ -225,6 +231,8 @@ def edit(ctx, id, todo_properties, interactive):
 
     if changes:
         todo.save()
+        if old_list.name != todo.list.name:
+            database.move(todo, todo.list, from_list=old_list)
         click.echo(ctx.obj['formatter'].detailed(todo))
     else:
         click.echo('No changes.')
@@ -277,7 +285,7 @@ def flush(ctx):
     '''
     database = ctx.obj['db']
     for todo in database.flush():
-        click.echo('Deleting {} ({})'.format(todo.uid, todo.summary))
+        click.echo(ctx.obj['formatter'].simple_action('Flushing', todo))
 
 
 @cli.command()
@@ -297,7 +305,7 @@ def delete(ctx, ids, yes):
         click.confirm('Do you want to delete those tasks?', abort=True)
 
     for todo in todos:
-        click.echo('Deleting {} ({})'.format(todo.uid, todo.summary))
+        click.echo(ctx.obj['formatter'].simple_action('Deleting', todo))
         ctx.obj['db'].delete(todo)
 
 
@@ -311,9 +319,7 @@ def copy(ctx, list, ids):
 
     for id in ids:
         todo = ctx.obj['db'].todo(id)
-        click.echo('Copying {} to {} ({})'.format(
-            todo.uid, list, todo.summary
-        ))
+        click.echo(ctx.obj['formatter'].compact(todo))
 
         ctx.obj['db'].save(todo, list)
 
@@ -328,9 +334,7 @@ def move(ctx, list, ids):
 
     for id in ids:
         todo = ctx.obj['db'].todo(id)
-        click.echo('Moving {} to {} ({})'.format(
-            todo.uid, list, todo.summary
-        ))
+        click.echo(ctx.obj['formatter'].compact(todo))
 
         ctx.obj['db'].move(todo, list)
 
@@ -342,7 +346,8 @@ def move(ctx, list, ids):
 @click.option('--location', help='Only show tasks with location containg TEXT')
 @click.option('--category', help='Only show tasks with category containg TEXT')
 @click.option('--grep', help='Only show tasks with message containg TEXT')
-@click.option('--sort', help='Sort tasks using these fields')
+@click.option('--sort', help='Sort tasks using these fields',
+              callback=_sort_callback)
 @click.option('--reverse/--no-reverse', default=True,
               help='Sort tasks in reverse order (see --sort). '
               'Defaults to true.')
@@ -355,9 +360,7 @@ def move(ctx, list, ids):
               help='Only show finished tasks')
 @click.option('--start', default=None, callback=_validate_start_date_param,
               help='Only shows tasks before/after given DATE')
-def list(
-        ctx, lists, all, location, category, grep, sort, reverse, due,
-        priority, start, done_only):
+def list(ctx, **kwargs):
     """
     List unfinished tasks.
 
@@ -372,21 +375,5 @@ def list(
     This is the default action when running `todo'.
     """
 
-    sort = sort.split(',') if sort else None
-
-    db = ctx.obj['db']
-    todos = db.todos(
-        due=due,
-        all=all,
-        category=category,
-        grep=grep,
-        lists=lists,
-        location=location,
-        reverse=reverse,
-        start=start,
-        sort=sort,
-        priority=priority,
-        complete=done_only,
-    )
-
+    todos = ctx.obj['db'].todos(**kwargs)
     click.echo(ctx.obj['formatter'].compact_multiple(todos))